refactor(control-plane): move quota should-run preparation into bounded module
